% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{change_native_status_code}
\alias{change_native_status_code}
\title{Change the native status code for a particular taxon at a particular site}
\usage{
change_native_status_code(df, taxon, site, new_code)
}
\arguments{
\item{df}{is the data frame returned by get_longform_cover}

\item{taxon}{is the taxonID column in the data frame}

\item{site}{is the identity of the NEON site (e.g. "JORN")}

\item{new_code}{is the NativeStatusCode value to change to}
}
\description{
Sometimes even though a particular species identity is not known, the end
user can still determine its native status. For example, maybe the taxon
was identified to the genus level, and the local flora confirms that all
plants in that genus are native at that particular site. This function
allows for post-hoc modification of the native status code for cases like this.
}
\examples{

# download the NEON plant diversity data for the Jornada Experimental Range
# raw_div<-download_plant_div(sites = "JORN")

# convert to longform cover
# lf_div <- get_longform_cover(raw_div)

# change all of the unknown Abutilon spp to native
# modified_lf_div <- change_native_status_code(lf_div, "ABUTI", "JORN", "N")

}
