% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D14}
\alias{D14}
\title{Plant Presence and Percent Cover Data for Domain 14}
\format{
## 'D14'
A list with 8 items, 2 of which are used by neonPlantEcology
}
\source{
<https://www.neonscience.org>
}
\usage{
D14
}
\description{
This includes Jornada Experimental Range and Santa Rita Experimental Range
}
\keyword{datasets}
