% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D14_release23}
\alias{D14_release23}
\title{Plant Presence and Percent Cover Data for Domain 14 (old release)}
\format{
## 'D14'
A list with 8 items, 2 of which are used by neonPlantEcology
}
\source{
<https://doi.org/10.48443/9579-a253>

<https://data.neonscience.org/data-products/DP1.10058.001/RELEASE-2023>
}
\usage{
D14_release23
}
\description{
This includes Jornada Experimental Range and Santa Rita Experimental Range
}
\keyword{datasets}
