% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{get_diversity_info}
\alias{get_diversity_info}
\title{Get plant biodiversity information for NEON plots}
\usage{
get_diversity_info(
  neon_div_object,
  scale = "plot",
  trace_cover = 0.5,
  fix_unks = FALSE,
  families = "Poaceae",
  spp = "Bromus tectorum"
)
}
\arguments{
\item{neon_div_object}{the raw vegan::diversity data downloaded using
neondiversity::download_plant_div() or #' the function
neonUtilities::loadByProduct() with the dpID arguement set to "DP1.10058.001".}

\item{scale}{what level of aggregation? This can be "1m", "10m", "100m", or "plot",
which is the default.}

\item{trace_cover}{cover value for subplots where only occupancy was recorded}

\item{fix_unks}{Should the unknown codes be altered with the "unk_fixer()"
function? Defaults to false. This requires manual investigation and editing
of the unk_fixer function.}

\item{families}{Which specific families should the metrics be calculated for?
This can be a concatenated vector if the user want more than one family.}

\item{species}{Which specific species should the metrics be calculated for?
This can be a concatenated vector if the user want more than one species.}
}
\description{
get_diversity_info calculates various biodiversity and cover indexes at the
plot or subplot scale at each timestep for each plot. Outputs a data frame
with number of species, percent cover, relative percent cover, and shannon
diversity, for natives, exotics and all species. Also calculates all of these
metrics for the families and/or species of your choice.
}
\examples{
x <- download_plant_div("SRER")
plot_level <- neondiversity::get_diversity_info(neon_div_object = x, scale = "plot")
}
