% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{get_longform_cover}
\alias{get_longform_cover}
\title{Convert raw NEON diversity object to longform plant cover data frame}
\usage{
get_longform_cover(
  neon_div_object,
  trace_cover = 0.5,
  scale = "plot",
  fix_unks = FALSE
)
}
\arguments{
\item{neon_div_object}{the raw diversity data downloaded using
neondiversity::download_plant_div() or the function
neonUtilities::loadByProduct() with the dpID arguement set to "DP1.10058.001".}

\item{trace_cover}{cover value for subplots where only occupancy was recorded}

\item{scale}{what level of aggregation? This can be "1m", "10m", "100m", or "plot",
which is the default.}

\item{fix_unks}{Should the unknown codes be altered with the "unk_fixer()"
function? Defaults to false. This requires manual investigation and editing
of the unk_fixer function.}
}
\description{
The diversity data from NEON comes as a list containing 2 data frames of data
that need to be combined, among other things. Here, we take those two data
frames and combine them into a longform data frame that can then be further
modified for analysis. Most of the unneccessary information from the raw data
has been removed. Column names that remain are plotID, subplotID, year,
taxonID, cover, scientificName, nativeStatusCode, family, and site.
}
\examples{
raw_div<-download_plant_div(sites = "SRER")
lf_div <- get_longform_cover(raw_div)
}
