% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{get_plot_info}
\alias{get_plot_info}
\title{Get plot information from a community matrix}
\usage{
get_plot_info(comm)
}
\arguments{
\item{comm}{the community matrix object created by get_community_matrix()}
}
\description{
The get_community_matrix() function is designed to work with the vegan
package, and one of the requirements of vegan functions is that there are
only numeric columns in community matrices. Therefore, all of the metatdata
is collapsed into the rownames. This function allows you to extract that
very basic metadata back out to a more easily interpretable data frame.
}
