% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{name_cleaner}
\alias{name_cleaner}
\title{Clean scientific names}
\usage{
name_cleaner(lf_cover)
}
\arguments{
\item{lf_cover}{the longform cover table created by using
neondiveRsity::get_longform_cover().}
}
\description{
This function does some straight forward grep-ing to aggregate to clean up
some of the latin names
}
