% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{npe_groundcover}
\alias{npe_groundcover}
\title{Get ground cover and other variables}
\usage{
npe_groundcover(neon_div_object, scale = "plot", timescale = "annual")
}
\arguments{
\item{neon_div_object}{the raw diversity data downloaded using
neonPlantEcology::download_plant_div() or the function
neonUtilities::loadByProduct() with the dpID arguement set to "DP1.10058.001".}

\item{scale}{the spatial scale of aggregation. Can be "1m", "10m", "100m",
"plot" or "site". default is "plot".}

\item{timescale}{The temporal scale of aggregation. Can be "all", "annual" or
"subannual" in the case of multiple sampling bouts per year. Defaults to "annual".}
}
\description{
Get ground cover and other variables
}
