% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{npe_name_cleaner}
\alias{npe_name_cleaner}
\title{Clean scientific names}
\usage{
npe_name_cleaner(lf_cover)
}
\arguments{
\item{lf_cover}{the longform cover table created by using}
}
\value{
a data frame
}
\description{
This function does some straight forward grep-ing to aggregate to clean up
some of the latin names
}
\examples{
data("D14")
npe_longform("D14") |> npe_name_cleaner()
}
