% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{npe_plot_centroids}
\alias{npe_plot_centroids}
\title{Download and join spatial information to a neonPlantEcology output data frame}
\usage{
npe_plot_centroids(
  df,
  type = "latlong",
  spatial_only = TRUE,
  input = "community_matrix"
)
}
\arguments{
\item{df}{a neonPlantEcology-produced data frame}

\item{type}{what type of ancillary data structure you want joined. Can be
"spatial", which will turn the data frame into an sf data frame, or "latlong",
which will add the latitudes and longitudes and other ancillary data as
columns only.}

\item{spatial_only}{set to TRUE if you only want the coordinates and none of
the ancillary variables.}

\item{input}{to what kind of neonPlantEcology product are you appending? Can
be "community_matrix", "longform_cover", or "summary_info".}
}
\value{
a data frame
}
\description{
Download and join spatial information to a neonPlantEcology output data frame
}
