% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{npe_species}
\alias{npe_species}
\title{get cover by species}
\usage{
npe_species(
  neon_div_object,
  scale = "plot",
  trace_cover = 0.5,
  dissolve_years = FALSE,
  name_cleaner = FALSE
)
}
\description{
This extracts the cover and relative cover of all species into a data frame.
This is basically the same as npe_community_matrix, except the species names
are full, it's in a tibble format, and the plot information is not collapsed
into the row names.
}
