% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{npe_update_subplots}
\alias{npe_update_subplots}
\title{Change subplot names}
\usage{
npe_update_subplots(neon_div_object)
}
\arguments{
\item{neon_div_object}{a list downloaded using npe_download containing Plant Presence and Percent Cover data}
}
\value{
a
}
\description{
The 2024 release features a change in subplot names. This function changes
subplot names of the Plant Presence and Percent Cover raw list object
from the old format to the new format, to ensure backwards
compatibility. This is mostly an internal helper function
}
\examples{
data("D14")
D14_updated <- npe_update_subplots(D14)
}
