% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_centroids}
\alias{plot_centroids}
\title{Plot centroids for the entire NEON network}
\format{
## 'plot_centroids'
A simple feature collection with 3842 features and 36 fields
}
\source{
<https://www.neonscience.org>
}
\usage{
plot_centroids
}
\description{
Plot centroids for the entire NEON network
}
\keyword{datasets}
