% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_polygons}
\alias{site_polygons}
\title{Polygons of National Ecological Observatory Network Core and Relocatable Terrestrial Sites}
\format{
## 'site_polygons'
A simple feature collection with 59 features and 8 fields
\describe{
 \item{domainNumb}{Domain Number}
 \item{domainName}{Domain Name}
 \item{siteType}{Site type. Core or Relocatable}
 \item{siteName}{Site name}
 \item{siteID}{Four letter site ID. Used in npe_download()}
 \item{siteHost}{Organization hosting the site}
 \item{areaKm2}{Area of the site in square kilometers}
 \item{acres}{Area of the site in acres}
 \item{geometry}{list column containing geometry information for each polygon}
}
}
\source{
<https://www.neonscience.org>
}
\usage{
site_polygons
}
\description{
Note: Some sites have more than one polygon. There are 59 polygons and 47 total sites.
}
\keyword{datasets}
