% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unk_investigation.R
\name{unk_fixer}
\alias{unk_fixer}
\title{Manually change unknown native status codes}
\usage{
unk_fixer(df)
}
\description{
Usually at any NEON site, there will be plenty of unknown plants for various
reasons. Often, it is possible to use the information that is available to
at least determine if a particular unknown species is native or introduced.
For example, it might be documented in the local flora that all plants of a
certain family or genus are native. This function allows the user to manually
change those codes. However it must be manually edited at this point.
}
