% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{vegify}
\alias{vegify}
\title{Create a species abundance or occurrence matrix}
\usage{
vegify(
  neon_div_object,
  scale = "plot",
  trace_cover = 0.5,
  fix_unks = FALSE,
  binary = FALSE
)
}
\arguments{
\item{neon_div_object}{the raw diversity data downloaded using
neondiversity::download_plant_div() or #' the function
neonUtilities::loadByProduct() with the dpID arguement set to "DP1.10058.001".}

\item{scale}{what level of aggregation? This can be "1m", "10m", "100m", or "plot",
which is the default.}

\item{trace_cover}{cover value for subplots where only occupancy was recorded}

\item{fix_unks}{Should the unknown codes be altered with the "unk_fixer()"
function? Defaults to false. This requires manual investigation and editing
of the unk_fixer function.}

\item{binary}{should the matrix be converted from percent cover to binary?}
}
\description{
vegify creates a wide matrix of species cover or binary (presence/absence)
values with the plot/subplot/year as rownames. This is useful for the vegan
package, hence the name.
}
