% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{download_plant_div}
\alias{download_plant_div}
\title{Plant diversity data downloader}
\usage{
download_plant_div(sites = "SRER")
}
\arguments{
\item{sites}{a vector of NEON site abbreviations. Defaults to "SRER"}
}
\description{
A function to download plant diversity data from the NEON API.
}
\details{
Data comes in two separate components:
1m2 subplots with cover estimates (first item of the list)
10 and 100m2 subplots with presence only (second item of the list)
}
\examples{
diversity_object <- download_plant_div(sites = "SRER")
}
\keyword{diversity}
\keyword{download}
\keyword{neon}
