% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{npe_plot_centroids}
\alias{npe_plot_centroids}
\title{Download and join spatial information to a neonPlantEcology output data frame}
\usage{
npe_plot_centroids(
  df,
  dest_dir = file.path(getwd(), "tmp"),
  type = "latlong",
  spatial_only = T,
  input = "community_matrix"
)
}
\arguments{
\item{df}{a neonPlantEcology-produced data frame}

\item{dest_dir}{where to download the files}

\item{type}{what type of ancillary data structure you want joined. Can be
"spatial", which will turn the data frame into an sf data frame, or "latlong",
which will add the latitudes and longitudes and other ancillary data as
columns only.}

\item{input}{to what kind of neonPlantEcology product are you appending? Can
be "community_matrix", "longform_cover", or "summary_info".}
}
\description{
Download and join spatial information to a neonPlantEcology output data frame
}
