% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_data_prep.R
\name{npe_site_ids}
\alias{npe_site_ids}
\title{get site ids}
\usage{
npe_site_ids(all = FALSE, domain = NA, type = NA)
}
\arguments{
\item{all}{if TRUE, returns a vector of all siteID codes}

\item{domain}{can be one or more domain codes, as a character vector.
e.g. domain = c("D01", "D14")}

\item{type}{can be "Core Terrestrial" or "Relocatable Terrestrial"}
}
\description{
This uses the site boundary shapefile (obtainable by data('sites')) to get a
list of siteID codes to feed into npe_download.
}
